Add support for conditional access to sections
Sections which are not currently visible to the student were being shown in the block, but would give an unfriendly error message stating that they don't have permission to view hidden sections if clicked.  This change makes the block check for availability - and if unavailable, whether it should show greyed out or not at all.